serverless-fastapi/api/requirements.txt
fastapi==0.61.1
mangum
requests>=2.24.0
uvicorn==0.12.2
jinja2
pyjwt
